requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Pillow==10.2.0
fpdf2==2.7.8
arabic-reshaper==3.0.0
python-bidi==0.4.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-Mail==0.9.1
flask-login==0.6.3
werkzeug==3.0.1
